+++ internal/provider/data_source_iosxe_access_list_extended_test.go
@@ -44,6 +44,7 @@ func TestAccDataSourceIosxeAccessListExtended(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_access_list_extended.test", "entries.0.syn", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_access_list_extended.test", "entries.0.urg", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_access_list_extended.test", "entries.0.dscp", "46"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_access_list_extended.test", "entries.0.log", "true"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -77,6 +78,7 @@ func testAccDataSourceIosxeAccessListExtendedConfig() string {
 	config += `		syn = true` + "\n"
 	config += `		urg = true` + "\n"
 	config += `		dscp = "46"` + "\n"
+	config += `		log = true` + "\n"
 	config += `	}]` + "\n"
 	config += `}` + "\n"