+++ internal/provider/resource_iosxe_access_list_extended_test.go
@@ -45,6 +45,7 @@ func TestAccIosxeAccessListExtended(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_extended.test", "entries.0.syn", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_extended.test", "entries.0.urg", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_extended.test", "entries.0.dscp", "46"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_extended.test", "entries.0.log", "true"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -93,6 +94,7 @@ func testAccIosxeAccessListExtendedConfig_all() string {
 	config += `		syn = true` + "\n"
 	config += `		urg = true` + "\n"
 	config += `		dscp = "46"` + "\n"
+	config += `		log = true` + "\n"
 	config += `	}]` + "\n"
 	config += `}` + "\n"
 	return config