+++ internal/provider/resource_iosxe_access_list_standard_test.go
@@ -32,6 +32,7 @@ func TestAccIosxeAccessListStandard(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_standard.test", "entries.0.remark", "Description"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_standard.test", "entries.0.deny_prefix", "10.0.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_standard.test", "entries.0.deny_prefix_mask", "0.0.0.255"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_access_list_standard.test", "entries.0.deny_log", "true"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -67,6 +68,7 @@ func testAccIosxeAccessListStandardConfig_all() string {
 	config += `		remark = "Description"` + "\n"
 	config += `		deny_prefix = "10.0.0.0"` + "\n"
 	config += `		deny_prefix_mask = "0.0.0.255"` + "\n"
+	config += `		deny_log = true` + "\n"
 	config += `	}]` + "\n"
 	config += `}` + "\n"
 	return config